create install_utils.py for better structure
